Refactor the code to remove isEndOfMessage added flag
When we started writing carbon-transport, a decision was taken to hide all the
Netty related stuff from application layer. However, later it was decided that
there is no need to hide netty stuff. This acutally allows us to write a more
simpler and readable code.

Doing so also allowed us to fix the streaming issue. Hence that fix also included
in this PR.